refactored instrumentatiion js.
separated to it's own file, which also required moving the loadScript
function to a utils file to avoid a circular import.

Also added automatic id get and instrumentation of blocks. Each blocks
is now auto-added to the data layer, and can set an optional
data-dataLayer attr on the div.block which will be added as the block
info.